+++ engine/servermaster/campaign_test.go
@@ -287,54 +287,3 @@ func TestLeaderLoopWatchLeader(t *testing.T) {
 	cancel()
 	wg.Wait()
 }
-
-func TestCampaignMeetLeaseExpire(t *testing.T) {
-	t.Parallel()
-
-	ctx, cancel := context.WithCancel(context.Background())
-	defer cancel()
-	name := "server-master-campaign-meet-lease-expire"
-	addr, etcd, client, cleanFn := test.PrepareEtcd(t, name)
-	defer cleanFn()
-
-	cfg := NewConfig()
-	cfg.Etcd.Name = name
-	cfg.AdvertiseAddr = addr
-	id := genServerMasterUUID(name)
-	s := &Server{
-		id:         id,
-		cfg:        cfg,
-		etcd:       etcd,
-		etcdClient: client,
-		info:       &model.NodeInfo{ID: model.DeployNodeID(name)},
-	}
-	preRPCHook := rpcutil.NewPreRPCHook[pb.MasterClient](
-		s.id,
-		&s.leader,
-		s.masterCli,
-		&s.leaderInitialized,
-		s.rpcLogRL,
-	)
-	s.masterRPCHook = preRPCHook
-
-	sessionCfg, err := s.generateSessionConfig()
-	require.Nil(t, err)
-	session, err := cluster.NewEtcdSession(ctx, client, sessionCfg)
-	require.Nil(t, err)
-	err = session.Reset(ctx)
-	require.Nil(t, err)
-	_, _, err = session.Campaign(ctx, time.Second)
-	require.Nil(t, err)
-
-	// simulate server lease expire, campaign will fail
-	// _, err = client.Revoke(ctx, session.(*cluster.EtcdSession).session.Lease())
-	// require.Nil(t, err)
-	// err = s.campaign(ctx, time.Second)
-	// require.Regexp(t, ".*etcdserver: requested lease not found", err)
-
-	// after handle campaign error, another try of campaign will success
-	// err = s.handleCampaignError(ctx, err)
-	// require.Nil(t, err)
-	// err = s.campaign(ctx, time.Second)
-	// require.Nil(t, err)
-}
